Revert "BUG: Fix bug with paths (#11639)"
This reverts commit 4eb9ef04c50710dba840ceaac6dc4f69ffefdd31.
