+++ tutorials/preprocessing/40_artifact_correction_ica.py
@@ -29,7 +29,6 @@ sample_data_folder = mne.datasets.sample.data_path()
 sample_data_raw_file = os.path.join(sample_data_folder, 'MEG', 'sample',
                                     'sample_audvis_filt-0-40_raw.fif')
 raw = mne.io.read_raw_fif(sample_data_raw_file)
-
 # Here we'll crop to 60 seconds and drop gradiometer channels for speed
 raw.crop(tmax=60.).pick_types(meg='mag', eeg=True, stim=True, eog=True)
 raw.load_data()
